크롤링 코드 리팩토링 및 도서 이미지 url 가져옴
