+++ crawling/__init__.py
@@ -1,0 +1,1 @@
+# 해당 디렉터리를 파이썬 패키지로 만들어줌
